pyproject.toml
[project]
name = "pm-kit"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["click", "pyyaml", "jinja2", "requests"]

[project.scripts]
pm-kit = "pm_kit.cli:main"

[tool.uv]
dev-dependencies = [
    "basedpyright>=1.39.0",
    "pytest",
    "ruff>=0.15.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
